Room Boundary System
You can now easily create a room by attaching the RoomBoundary script
and connecting the room to other room's through the editor. When you
reach the edges of a room boundary the camera will stop moving so that
you can't see into the next room. If the player moves into another room,
the camera will automatically readjust itself to the boundary of the new
room.