Fixed Vertical Camera Movement
After doing a test on vertical movement of the camera I found that it
wasn't working as intended, after spending a bunch of time wondering how
my code was wrong I finally realised that I had set the camera to have
an offset on the y axis causing the camera based calculations to appear
wrong. Setting the offset in Unity back to 0 fixed the problem.