[DONE] Camera Fade In / Out
Camera fade now works correctly when transitioning rooms. The camera is
intended to replicate the camera seen in Castlevania: Symphony of the
Night, after watching the fade effect dozens of times in slow motion I
realised that the reason my fade effect looked wrong was because the
camera's position was changing to the next room before the screen had
fully faded to black, this meant that you could see the camera's jarring
teleporting as it was fading to black. This is now corrected and the
camera does not change rooms until the screen is completely black
preventing the player from seeing the transition, making the transition
feel much smoother.