Nothing Works
Attempting to combine my camera / room systems with Ryan's physics /
nanobots systems. Proving to be a weird one, Unity seems to be causing
scripts to behave differently in different scenes. My scripts work fine
in my scene, but when using those exact same scripts in another scene
they seem to break for no reason. After a bit of debugging it seems the
variables are not working across scenes. For example I have a boolean
that shows if the camera is currently transitioning between rooms or
not, it is false at the start, but in this new scene it is true at the
start. A lot of weird behaviour, this commit will probably be deleted in
the future once we get this figured out.