Max's Camera Easing
Max continues to doubt my camera creating skills. We have implemented a
slight variation of the camera that handles the 'easing' effect
differently, the problem with the current iteration is that the speed of
the camera keeps changing as the distance between it and the player
change, causing jarring movements with the camera.