More Rooms
Previously a room could have four rooms linked to it, one above, below,
left and right. These links allowed for the camera to correctly update
itself whenever the player transitioned rooms. Now, you are able to add
multiple rooms to the same side, so for example you may want to link 2
different rooms to the left if it is a tall room with mutliple floors.