First Public Build
This will be the first build of the project that is available to all our
group members. Ryan and I have worked together to get my camera / room
system and his player / bots working correctly. I have taken this and
reworked the entire project so that there is an example scene that
already has the required scripts / objects in place for the level
designers to begin working. Each person in the group will have their own
scene where they will be adding their content until they can get it
working. Then there is a master scene and area for finalised assets that
will be needed further down the line when we are bringing everyone's
levels together.