Bot Capsule
Implemented a basic version of the bot capsule, there are two scripts,
one defines the pick-up functionality, at the moment it just destroys
the game object on trigger however the Update() function will feature
some animation later on. There is also a create object script that
determines which object we create and how many of them to create.