Fixed Cattle Grid and Grate
The meshes used for the Cattle Grid and Grate have been updated so that
they work as intended. Changed the default values for the player's jump
and movement velocity, this should remove any confusion with people
thinking the player has broken when really, it's just that the speed has
reset back to 0 preventing movement.