Bot Counter
The bot counter is not functional. It has a progress bar that represents
how many bots the player has on a scale of 0 to 100. There is also a
number value that shows the exact amount of bots the player has. There
are still some things to fix and account for and the image is too high a
quality for the UI so we are waiting on an updated image for the
progress bar.