Organized scripts and Updated Weighted Plates
Moved all scripts to do with objects in the world into its own folder
and subfolders (for each object). Weighted Plates can now detect the
number of bots within 1 ZU of it (the block surrounding the weighted
plate, size of all blocks). Boolean check on if it uses this additional
check.